pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "joblib>=1.4.2",
    "numpy>=2.2.5",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "flask-mail>=0.10.0",
    "python-dotenv>=1.1.0",
    "sendgrid>=6.11.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "beautifulsoup4>=4.13.4",
    "openai>=1.76.0",
    "google-generativeai>=0.8.5",
    "anthropic>=0.50.0",
    "trafilatura>=2.0.0",
    "pillow>=11.2.1",
    "nltk>=3.9.1",
    "textblob>=0.19.0",
    "langdetect>=1.0.9",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "gdown>=5.2.0",
    "razorpay>=1.4.2",
    "pytz>=2025.2",
    "google-genai>=1.25.0",
    "sift-stack-py>=0.7.0",
    "psutil>=7.0.0",
    "redis>=6.2.0",
    "opencv-python>=4.12.0.88",
    "mediapipe>=0.10.14",
    "scipy>=1.15.2",
    "twilio>=9.7.1",
]
